requirements.txt
certifi==2018.8.24
chardet==3.0.4
gitdb2==2.0.4
GitPython==2.1.11
idna==2.7
Jinja2==2.10
MarkupSafe==1.0
mistune==0.8.3
mistune-contrib==0.1
requests==2.19.1
smmap2==2.0.4
urllib3==1.23
